[binding] Refactor: move the code for "all bindings" to a separate class.
