+++ to.etc.domui/src/to/etc/domui/dom/html/NodeBase.java
@@ -1289,11 +1289,11 @@ abstract public class NodeBase extends CssBase implements INodeErrorDelegate {
 		if(null != el) {
 			sb.append(":").append(el);
 		}
-		SimpleBinder binding = SimpleBinder.findBinding(this, "value");
+		SimpleBinder binding = DefaultBindingHandler.findBinding(this, "value");
 		if(binding != null) {
 			sb.append(" ").append(binding);
 		} else {
-			binding = SimpleBinder.findBinding(this, "bindValue");
+			binding = DefaultBindingHandler.findBinding(this, "bindValue");
 			if(binding != null) {
 				sb.append(" ").append(binding);
 			}
@@ -1845,7 +1845,7 @@ abstract public class NodeBase extends CssBase implements INodeErrorDelegate {
 
 	@Nonnull
 	public List<UIMessage> getBindingErrors() throws Exception {
-		return SimpleBinder.getBindingErrors(this);
+		return DefaultBindingHandler.getBindingErrors(this);
 	}
 
 	/**
@@ -1854,7 +1854,7 @@ abstract public class NodeBase extends CssBase implements INodeErrorDelegate {
 	 * @return
 	 */
 	public boolean bindErrors() throws Exception {
-		return SimpleBinder.reportBindingErrors(this);
+		return DefaultBindingHandler.reportBindingErrors(this);
 	}
 
 	/**
